fix: changing amount type to integer

+++ app/models/tables/purchases.py
@@ -7,4 +7,4 @@ class Purchases(Base):
     __tablename__: str = "purchases"
 
     company_code: Mapped[str] = mapped_column()
-    amount: Mapped[float] = mapped_column()
+    amount: Mapped[int] = mapped_column()
